Add the rag feature code
