+++ rag/embedding_function.py
@@ -1,0 +1,30 @@
+from chromadb import Documents, EmbeddingFunction, Embeddings
+from google.api_core import retry
+from google import genai
+from google.genai import types
+
+# Define a helper to retry when per-minute quota is reached.
+is_retriable = lambda e: (isinstance(e, genai.errors.APIError) and e.code in {429, 503})
+
+class GeminiEmbeddingFunction(EmbeddingFunction):
+    # Specify whether to generate embeddings for documents, or queries
+    document_mode = True  
+    def __init__(self, client):
+        self.client = client
+
+    @retry.Retry(predicate=is_retriable)
+    def __call__(self, input: Documents) -> Embeddings:
+        if self.document_mode:
+            embedding_task = "retrieval_document"
+        else:
+            embedding_task = "retrieval_query"
+
+        response = self.client.models.embed_content(
+            model="models/text-embedding-004",
+            contents=input,
+            config=types.EmbedContentConfig(
+                task_type=embedding_task,
+            ),
+        )
+        return [e.values for e in response.embeddings]
+
